+++ src/components/Chips/Chips.tsx
@@ -14,7 +14,6 @@ import { SubIcon, VibeComponent, VibeComponentProps } from "../../types";
 import useHover from "../../hooks/useHover";
 import useSetFocus from "../../hooks/useSetFocus";
 import { ComponentDefaultTestId } from "../../tests/constants";
-import { backwardCompatibilityForProperties } from "../../helpers/backwardCompatibilityForProperties";
 import useClickableProps from "../../hooks/useClickableProps";
 import { BEMClass } from "../../helpers/bem-helper";
 import "../Clickable/Clickable.scss";
@@ -66,13 +65,16 @@ interface ChipsProps extends VibeComponentProps {
    */
   ariaLabel?: string;
   /**
-   * Should element be focusable & clickable - for backward compatability
+   * Deprecated, there is no need to use this prop for implementing clickable chips. Please use onClick for this purpose.
+   * @deprecated
    */
   clickable?: boolean;
   /**
-   * Backward compatibility for props naming - please use clickable instead
+   * Deprecated, there is no need to use this prop for implementing clickable chips. Please use onClick for this purpose.
+   * @deprecated
    */
   isClickable?: boolean;
+  disableClickableBehavior?: boolean;
 }
 
 const Chips: VibeComponent<ChipsProps, HTMLElement> & {
@@ -100,15 +102,13 @@ const Chips: VibeComponent<ChipsProps, HTMLElement> & {
       onClick,
       noAnimation = false,
       ariaLabel,
-      isClickable,
-      clickable,
-      dataTestId
+      dataTestId,
+      disableClickableBehavior = false
     },
     ref
   ) => {
-    const overrideClickable = backwardCompatibilityForProperties([clickable, isClickable], false);
     const overrideDataTestId = dataTestId || getTestId(ComponentDefaultTestId.CHIP, id);
-    const hasClickableWrapper = overrideClickable && (!!onClick || !!onMouseDown);
+    const hasClickableWrapper = (!!onClick || !!onMouseDown) && !disableClickableBehavior;
     const hasCloseButton = !readOnly && !disabled;
 
     const focusRef = useRef(null);
@@ -143,7 +143,7 @@ const Chips: VibeComponent<ChipsProps, HTMLElement> & {
 
     const onClickCallback = useCallback(
       (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
-        if ((e.target as HTMLElement) !== iconButtonRef.current && onClick) {
+        if (onClick !== undefined && (e.target as HTMLElement) !== iconButtonRef.current) {
           e.preventDefault();
           onClick(e);
         }
@@ -184,6 +184,8 @@ const Chips: VibeComponent<ChipsProps, HTMLElement> & {
         }
       : {
           ref: mergedRef,
+          onClick: onClickCallback,
+          onMouseDown,
           id: id,
           "data-testid": overrideDataTestId,
           className: overrideClassName,