+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## 1.96.0 (2023-02-13)
+
+#### Bug Fixes
+* [#1105](https://github.com/mondaycom/monday-ui-react-core/pull/1105) Set chips clickable behavior as default instead of extra prop for backward support and clearer API  ([@hadasfa](https://github.com/hadasfa))
+
+#### New Features
+* [#1103](https://github.com/mondaycom/monday-ui-react-core/pull/1103) Tooltip.hideShowTriggers - static property ([@SergeyRoyt](https://github.com/SergeyRoyt))
+
+#### Documentation
+* [#1099](https://github.com/mondaycom/monday-ui-react-core/pull/1099) Feature/sergeyro/remove storybook styles overrides ([@SergeyRoyt](https://github.com/SergeyRoyt))
+
+## 1.95.3 (2023-02-08)
+
+#### Bug Fixes
+* [#1100](https://github.com/mondaycom/monday-ui-react-core/pull/1100) Fix modal zIndex breaking change ([@SergeyRoyt](https://github.com/SergeyRoyt))
+
+#### Documentation
+* [#1097](https://github.com/mondaycom/monday-ui-react-core/pull/1097) Add founding designers to welcome page ([@hadasfa](https://github.com/hadasfa))
+
+## 1.95.2 (2023-02-07)
+
+#### Bug Fixes
+* [#1098](https://github.com/mondaycom/monday-ui-react-core/pull/1098) Fix modal title font family ([@orrgottlieb](https://github.com/orrgottlieb))
+
+## 1.95.1 (2023-02-06)
+
+#### Bug Fixes
+* [#1096](https://github.com/mondaycom/monday-ui-react-core/pull/1096) Feature/sergeyro/fix modal z index prop ([@SergeyRoyt](https://github.com/SergeyRoyt))
+
 ## 1.95.0 (2023-02-06)
 
 #### New Features